ActivityPanel component refactored
